+++ CHANGELOG.md
@@ -64,6 +64,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * `extractInnerEdges()` to `PGS_Meshing`. Extracts all inner edges from a mesh.
 * `centerLine()` to `PGS_Contour`. Determines the longest center line passing through a given shape.
 * Additional signatures for `PGS_Conversion.toWKB()` and `.fromWKB()` that write/read the binary shape representation into a file.
+* `pointOnExteriorByDistance()` to `PGS_Processing`. Extracts a point from the perimeter (exterior) of the given shape at some distance along its perimeter.
 
 ### Changed
 * Reimplemented `PGS_Processing.equalParition()`. New algorithm is ~2x faster. Also removed `precise` parameter from method signature (no longer necessary).